Pin Node to a specific version
**Why**: The mymove repo uses a specific version of Node, as specified
in the `.node-version` file. The mymove repo also has a
`check-node-version` script that ensures that the local environment has
that specific version installed. Previously, that script was more lax
than it should be. It allowed any 14.x version, but now we updated it
to check for the exact version. This ensures all developers are all
using the same version and eliminates any issues that might arise from
version mismatches.

Similarly, we want all Docker images and the deployed environments to
use the exact same version of Node. Previously, this Docker image was
grabbing the latest 14.x Node version whenever the image was updated,
which caused it to be ahead of the version used in the mymove repo.

This PR now installs a specific Node version. As of today, it's 14.17.1.
If we decide to update the mymove repo in the future, we'll need to
update this image with that same version.