Adding build config that builds ARM in CircleCI;
This build config actually builds ARM in CircleCI. It doesn't do
anything besides print out the `uname -m` command. But this is good
enough for now and separates the code paths so that this can live on in
the repository without interfering with the rest of the x86_64
architecture that's required by the application.