Automating the test-branch-name job & docs;
Updating a number of things here. The test-branch-name is the largest
change. It's now using the folders found in the repository to determine
what the tags are that it should avoid. I've also added the
`milmove-atlantis` tag to the ignorance for the `build` job as well.
Lastly, I've added some documentation that discusses some of the CI/CD
updates so folks understand what all the above is doing for them.